+++ xen/common/libacpi/build.c
@@ -480,7 +480,7 @@ static int new_vm_gid(struct acpi_config *config)
     return 1;
 }
 
-void acpi_build_tables(struct acpi_config *config, unsigned int physical)
+void acpi_build_tables(struct acpi_config *config, unsigned long physical)
 {
     struct acpi_20_rsdp *rsdp;
     struct acpi_20_rsdt *rsdt;
